refactor: :art: Create new personalized hooks and utils
