+++ src/components/CityList/CityList.jsx
@@ -1,15 +1,14 @@
-import React, { useState, useEffect } from 'react'
+import React from 'react'
 import PropTypes from 'prop-types'
-import axios from 'axios'
-import convertUnits from 'convert-units'
 import Alert from '@material-ui/lab/Alert'
 import Grid from '@material-ui/core/Grid'
 import List from '@material-ui/core/List'
 import ListItem from '@material-ui/core/ListItem'
+import useCityList from '../../hooks/useCityList'
 import CityInfo from '../CityInfo'
 import Weather from '../Weather'
+import { getCityCode } from '../../utils/utils'
 
-const getCityCode = (city, countryCode) => `${city}-${countryCode}`
 // li: es un item (segun tag html)
 // renderCityAndCountry se va a convertir en una función que retorna otra función
 const renderCityAndCountry = eventOnClickCity => (cityAndCountry, weather) => {
@@ -45,43 +44,7 @@ const renderCityAndCountry = eventOnClickCity => (cityAndCountry, weather) => {
 // cities: es un array, y en cada item tiene que tener la ciudad, pero ademas el country
 // ul: tag html para listas no ordenadas
 const CityList = ({ cities, onClickCity }) => {
-    const [allWeather, setAllWeather] = useState({})
-    const [error, setError] = useState(null)
-
-    useEffect(() => {
-        const setWeather = async (city, countryCode) => {
-            const appid = "fdb503f509462933235611828b891ff4"
-            const url = `https://api.openweathermap.org/data/2.5/weather?q=${city},${countryCode}&appid=${appid}`
-
-            try {
-                const response = await axios.get(url)
-
-                const { data } = response
-                const temperature = Number(convertUnits(data.main.temp).from("K").to("C").toFixed(0))
-                const state = data.weather[0].main.toLowerCase()
-                const propName = getCityCode(city, countryCode)
-                const propValue = { temperature, state } // Ej:  { temperature: 10, state: "sunny" }
-
-                setAllWeather(allWeather => ({
-                    ...allWeather,
-                    [propName] : propValue
-                }))
-            } catch (error) {
-                if (error.response) {
-                    setError("Ha ocurrido un error en el servidor del clima")
-                } else if (error.request) {
-                    setError("Verifique la conexión a internet")
-                } else {
-                    setError("Error al cargar los datos")
-                }
-            }
-        }
-
-        cities.forEach(({ city, countryCode }) => {
-            setWeather(city, countryCode)
-        })
-
-    }, [cities])
+    const [allWeather, error, setError] = useCityList(cities)
 
     return (
         <div>
