+++ src/components/WelcomeScreen/WelcomeScreen.jsx
@@ -1,28 +1,10 @@
-import React, { useRef, useEffect, useState } from 'react'
+import React from 'react'
 import PropTypes from 'prop-types'
-import Clouds from 'vanta/dist/vanta.clouds.min'
-import * as THREE from 'three'
+import useVanta from '../../hooks/useVanta'
 
 const WelcomeScreen = ({ children }) => {
-    const myRefDiv = useRef(null)
-    const [vanta, setVanta] = useState(0)
 
-    useEffect(() => {
-        if (!vanta) {
-            setVanta(
-                Clouds({
-                    THREE,
-                    el: myRefDiv.current
-                })
-            )
-        }
-        // Al salir de la pantalla vamos a detener el efecto
-        return () => {
-            if (vanta) {
-                vanta.destroy()
-            }
-        }
-    }, [vanta])
+    const myRefDiv = useVanta()
 
     return (
         <div className="full" ref={myRefDiv}>
