+++ src/pages/CityPage.jsx
@@ -1,118 +1,15 @@
-import React, { useState, useEffect } from 'react'
-import axios from 'axios'
-import convertUnits from 'convert-units'
+import React from 'react'
 import Grid from '@material-ui/core/Grid'
-import moment from 'moment'
-import 'moment/locale/es'
-import { useParams } from 'react-router-dom'
 import CityInfo from './../components/CityInfo'
 import Weather from './../components/Weather'
 import WeatherDetails from './../components/WeatherDetails'
 import ForecastChart from './../components/ForecastChart'
 import Forecast from './../components/Forecast'
 import AppFrame from '../components/AppFrame'
-
-const dataExample = [
-    {
-        "dayHour": "Jue 18",
-        "min": 14,
-        "max": 22,
-    },
-    {
-        "dayHour": "Vie 06",
-        "min": 18,
-        "max": 27,
-    },
-    {
-        "dayHour": "Vie 12",
-        "min": 18,
-        "max": 28,
-    },
-    {
-        "dayHour": "Vie 18",
-        "min": 18,
-        "max": 25,
-    },
-    {
-        "dayHour": "Sab 06",
-        "min": 15,
-        "max": 22,
-    },
-    {
-        "dayHour": "Sab 12",
-        "min": 12,
-        "max": 19,
-    }
-]
-
-const forecastItemListExample = [
-    { hour: 18, state:"clear", temperature:17, weekDay:"Jueves" },
-    { hour: 6, state:"clouds", temperature:18, weekDay:"Viernes" },
-    { hour: 12, state:"drizzle", temperature:18, weekDay:"Sábado" },
-    { hour: 18, state:"clouds", temperature:19, weekDay:"Domingo" },
-    { hour: 14, state:"rain", temperature:17, weekDay:"Lunes" },
-    { hour: 10, state:"rain", temperature:17, weekDay:"Martes" },
-]
+import useCityPage from '../hooks/useCityPage'
 
 const CityPage = props => {
-
-    const [data, setData] = useState(null)
-    const [forecastItemList, setForecastItemList] = useState(null)
-
-    const { city, countryCode } = useParams()
-
-    useEffect(() => {
-        const getForecast = async () =>{
-            const appid = "fdb503f509462933235611828b891ff4"
-            const url = `https://api.openweathermap.org/data/2.5/forecast?q=${city},${countryCode}&appid=${appid}`
-
-            try {
-                const { data } = await axios.get(url)
-
-                const toCelsius = (temp) => Number(convertUnits(temp).from('K').to('C').toFixed(0))
-
-                const daysAhead = [0, 1, 2, 3, 4, 5]
-                const days = daysAhead.map(d => moment().add(d, 'd'))
-                const dataAux = days.map(day => {
-                    const tempObjArray = data.list.filter(item => {
-                        const dayOfYear = moment.unix(item.dt).dayOfYear()
-                        return dayOfYear === day.dayOfYear()
-                    })
-
-                    const temps = tempObjArray.map(item => item.main.temp)
-
-                    return ({
-                        dayHour: day.format('ddd'),
-                        min: toCelsius(Math.min(...temps)),
-                        max: toCelsius(Math.max(...temps)),
-                    })
-                })
-
-                setData(dataAux)
-
-                const interval = [4, 8, 12, 16, 20, 24]
-
-                const forecastItemListAux = data.list
-                    .filter((item, index) => interval.includes(index))
-                    .map(item => {
-                        return ({
-                            key: (item.dt),
-                            hour: moment.unix(item.dt).format('HH:mm'),
-                            weekDay: moment.unix(item.dt).format('dddd'),
-                            state: item.weather[0].main.toLowerCase(),
-                            temperature: toCelsius(item.main.temp)
-                        })
-                    })
-
-                setForecastItemList(forecastItemListAux)
-            } catch (error) {
-                console.log(error)
-            }
-        }
-        
-        getForecast()
-    }, [ city, countryCode ])
-    
+    const [city, chartData, forecastItemList] = useCityPage()
 
     const country = "España"
     const state = "clear"
@@ -132,7 +29,7 @@ const CityPage = props => {
                 </Grid>
                 <Grid item>
                     {
-                        data && <ForecastChart data={data}/>
+                        chartData && <ForecastChart data={chartData}/>
                     }
                 </Grid>
                 <Grid item>
